RPI 64 bit

+++ SBC/rpi-4/device_driver/firmware/radioberry_ioctl.h
@@ -14,7 +14,7 @@ struct rb_info_arg_t
 	
 	int fpga;
 	
-	float version;
+	int version;
 	
 	int rb_command;
 	int command;
